finishing implement api

+++ app/src/main/java/com/napa/foodstorechallengech3/presentation/feature/detailmenu/DetailMenuViewModel.kt
@@ -11,13 +11,14 @@ import com.napa.foodstorechallengech3.utils.ResultWrapper
 import kotlinx.coroutines.launch
 
 class DetailMenuViewModel(
-    private val extras: Bundle?, private val cartRepository: CartRepository
+    private val extras: Bundle?,
+    private val cartRepository: CartRepository
 ) : ViewModel() {
 
     val menu = extras?.getParcelable<Menu>(DetailMenuActivity.EXTRA_PRODUCT)
 
-    val priceLiveData = MutableLiveData<Double>().apply {
-        postValue(0.0)
+    val priceLiveData = MutableLiveData<Int>().apply {
+        postValue(0)
     }
     val menuCountLiveData = MutableLiveData<Int>().apply {
         postValue(0)
@@ -38,14 +39,14 @@ class DetailMenuViewModel(
     fun add() {
         val count = (menuCountLiveData.value ?: 0) + 1
         menuCountLiveData.postValue(count)
-        priceLiveData.postValue(menu?.price?.times(count) ?: 0.0)
+        priceLiveData.postValue(menu?.price?.times(count) ?: 0)
     }
 
     fun minus() {
         if((menuCountLiveData.value ?: 0) > 0){
             val count = (menuCountLiveData.value ?: 0) -1
             menuCountLiveData.postValue(count)
-            priceLiveData.postValue(menu?.price?.times(count) ?: 0.0)
+            priceLiveData.postValue(menu?.price?.times(count) ?: 0)
         }
     }
